The .ghc.env fix in ghc went in at the end of August
Change the minimum version we use to decide if ghc supports .ghc.env
files. Previously we declared that it required 8.0.2, but 8.0.2 is not
out yet so this makes things hard to test.

It was fixed in the 8.0.x branch at the end of August, so if ghc
declares itself to be 8.0.1.$date from September or later, then it's ok.